+++ CHANGELOG.md
@@ -6,13 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* Use ajv and a schema instead of runtypes for validation
+### Removed
+* runtypes dependency
 
 ## [4.3.0](https://github.com/animavirtuality/onlybots-core/compare/4.2.0...4.3.0) - February 1, 2023
 ### Changed
 * Remove `background_color` from metadata
 
 ## [4.2.0](https://github.com/animavirtuality/onlybots-core/compare/4.1.0...4.2.0) - January 27, 2023
-🔥
 ### Changed
 * Add `external_url` and `background_color` to metadata